+++ .env
@@ -1,0 +1,3 @@
+USERNAME = "postgres"
+PASSWORD = "password"
+JWT_SECRET_KEY = "somesecret"
