+++ app.py
@@ -1,16 +1,37 @@
-import config
 import json
 import jwt
+import os
 from flask import Flask, request
-from config import db, SECRET_KEY
-from modals.Drivers import Users
-from modals.Riders import Riders
+from flask_sqlalchemy import SQLAlchemy
+from flask_bcrypt import Bcrypt
+from dotenv import load_dotenv
+
+# os, dotenv, and load_dotenv() are what we need to use .env to hide confidential code
+load_dotenv()
+USERNAME = os.getenv("USERNAME")
+PASSWORD = os.getenv("PASSWORD")
+JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY")
 
 
 app = Flask(__name__)
-# make app available to fll files within entire folder
-app.config.from_object(config)
-db.init_app(app)
+app.config[
+    "SQLALCHEMY_DATABASE_URI"
+] = f"postgresql://{USERNAME}:{PASSWORD}@localhost/smartbus_backend"
+db = SQLAlchemy(app)
+
+bcrypt = Bcrypt()
+
+
+class Drivers(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(100))
+    password = db.Column(db.String(100))
+
+
+class Riders(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(100))
+    password = db.Column(db.String(100))
 
 
 with app.app_context():
@@ -22,17 +43,17 @@ def driverlogin():
     data = json.loads(request.get_data())
     if not data["username"] or not data["password"]:
         return "Please enter the required fields", 400
-    driver_login = Users.query.filter_by(username=data["username"]).first()
+
+    driver_login = Drivers.query.filter_by(username=data["username"]).first()
     hashed_password = driver_login.password
+
     # if not bcrypt.check_password_hash(hashed_password, data['password']):
     #   return "password you enter is not correct", 400
     # else:
     encoded_jwt = jwt.encode(
-        {"username": data["username"]}, SECRET_KEY, algorithm="HS256"
+        {"username": data["username"]}, JWT_SECRET_KEY, algorithm="HS256"
     )
-    # print(encoded_jwt)
     return encoded_jwt
-    # return data
 
 
 @app.route("/riderLogin", methods=["POST"])
@@ -41,10 +62,9 @@ def riderLogin():
     if not data["username"] or not data["password"]:
         return "Please enter the required fields", 400
     rider_login = Riders.query.filter_by(username=data["username"]).first()
-    print(rider_login)
     hashed_password = rider_login.password
     encoded_jwt = jwt.encode(
-        {"username": data["username"]}, SECRET_KEY, algorithm="HS256"
+        {"username": data["username"]}, JWT_SECRET_KEY, algorithm="HS256"
     )
     return encoded_jwt
 
@@ -58,16 +78,12 @@ def driveroutes():
     auth_token = request.headers.get("Authorization")
 
     try:
-        driver_entered_data = jwt.decode(auth_token, SECRET_KEY, algorithms="HS256")
-        current_logged_in_user = Users.query.filter_by(
+        driver_entered_data = jwt.decode(auth_token, JWT_SECRET_KEY, algorithms="HS256")
+        current_logged_in_user = Drivers.query.filter_by(
             username=driver_entered_data["username"]
         ).first()
 
-        print(current_logged_in_user.username)
-
         return "some route info", 200
-
-        # return jsonify(current_logged_in_user)
 
     except:
         return "Token is invalid", 401
