+++ env.example
@@ -1,0 +1,2 @@
+USERNAME = "your username"
+PASSWORD = "your password"
