--- modals/Drivers.py
@@ -1,7 +1,0 @@
-from config import db
-
-
-class Users(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(100))
-    password = db.Column(db.String(100))
